Fix problem_51.py
Unfortunately, the algorithm you proposed is incorrect. I ran it 100000 times and determined the probability of drawing each card at is original position so card[0] at the the index 0. The resulting average probability was around `0.03` and not the desired `1/52 ≈ 0.0192`. Therefore, the algorithm fails to ensure that each permutation of the deck is equally likely.

The algorithm should follow the following procedure:
1. Randomly pick one card out of 52 (52 possible ways to do this) and swap the picked card with the card at the end of the array 
2. Randomly pick one card out of 51 (51 possible ways to do this) and swap this picked card with the second last element in the array
3. ...
Total combinations are then 52!